Ingest preview

+++ src/main/kotlin/au/gov/api/ingest/APIController.kt
@@ -1,9 +1,6 @@
 package au.gov.api.ingest
 
 import au.gov.api.config.Config
-import au.gov.api.ingest.preview.DataImpl
-import au.gov.api.ingest.preview.EngineImpl
-import au.gov.api.ingest.preview.IngestImpl
 import com.beust.klaxon.JsonObject
 import com.beust.klaxon.Klaxon
 import com.beust.klaxon.Parser
@@ -11,16 +8,14 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import khttp.get
 import khttp.structures.authorization.BasicAuthorization
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.context.annotation.ClassPathScanningCandidateComponentProvider
 import org.springframework.core.env.Environment
-import org.springframework.core.type.filter.AnnotationTypeFilter
 import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.*
 import java.io.File
+import java.security.InvalidKeyException
 import java.security.MessageDigest
 import java.util.*
 import javax.servlet.http.HttpServletRequest
-import kotlin.collections.HashMap
 
 @RestController
 class APIController {
@@ -43,23 +38,19 @@ class APIController {
     fun setManifest(request: HttpServletRequest, @RequestBody mf: Manifest,
                     @RequestParam(required = false, defaultValue = "false") preview: Boolean): String {
         if (isAuthorisedToSaveService(request, mf.metadata.features.space!!)) {
-            if (!preview) {
-
-                if (mf.metadata.id.isNullOrBlank() && !isManifestUnique(mf)) return "This resource already exists"
-
-                var pipeOutputs = ExecuteManifest(mf)
-                if (mf.metadata.id.isNullOrBlank()) {
-                    mf.metadata.id = pipeOutputs.first().toString()
-                }
-                repository.save(mf)
-                logEvent(request, "Posted", "Manifest", mf.metadata.name!!, "Posted", ObjectMapper().writeValueAsString(mf))
-
-                return ObjectMapper().writeValueAsString(pipeOutputs)
-
-            } else {
-                //previewManifest(mf)
-                return "Not implimented"
+            if (preview) {
+                mf.assets.forEach { it.type = "${it.type}_preview" }
             }
+            if (mf.metadata.id.isNullOrBlank() && !isManifestUnique(mf)) return "This resource already exists"
+
+            var pipeOutputs = ExecuteManifest(mf)
+            if (mf.metadata.id.isNullOrBlank()) {
+                mf.metadata.id = pipeOutputs.first().toString()
+            }
+            repository.save(mf)
+            logEvent(request, "Posted", "Manifest", mf.metadata.name!!, "Posted", ObjectMapper().writeValueAsString(mf))
+
+            return ObjectMapper().writeValueAsString(pipeOutputs)
         } else {
             throw Unauthorised()
         }
@@ -69,32 +60,41 @@ class APIController {
     @ResponseStatus(HttpStatus.OK)
     @PostMapping("/upload")
     fun uploadFile(request: HttpServletRequest,
-                    @RequestParam(required = false, defaultValue = "") id: String,
-                    @RequestParam fileName: String,
-                    @RequestParam MIMEType: String,
-                    @RequestBody x:ByteArray): String {
-        var md5= hashString("MD5",x.toString())
-        val key = getKeyFromRequest(request)
-        if (id!="" && repository.findFileByKey(key).any { it.id.equals(id) }) {
-            repository.deleteFile(id)
-            md5=id
+                   @RequestParam(required = false, defaultValue = "") id: String,
+                   @RequestParam fileName: String,
+                   @RequestParam MIMEType: String,
+                   @RequestBody x: ByteArray): String {
+        if (isKeyValid(request)) {
+            var md5 = hashString("MD5", x.toString())
+            val key = getKeyFromRequest(request)
+            if (id != "" && repository.findFileByKey(key).any { it.id.equals(id) }) {
+                repository.deleteFile(id)
+                md5 = id
+            }
+            var base64file = String(java.util.Base64.getEncoder().encode(x), Charsets.UTF_8)
+            repository.saveFile(md5, fileName, MIMEType, key, base64file)
+            return md5
+        } else {
+            throw InvalidKeyException()
         }
-        var base64file = String(java.util.Base64.getEncoder().encode(x),Charsets.UTF_8)
-        repository.saveFile(md5,fileName,MIMEType,key,base64file)
-        return md5
     }
 
     @CrossOrigin
     @ResponseStatus(HttpStatus.OK)
     @DeleteMapping("/upload")
-    fun deleteFile(request: HttpServletRequest, @RequestParam id: String) : Boolean{
-        val key = getKeyFromRequest(request)
-        var success = false
-        if (repository.findFileByKey(key).any { it.id.equals(id) }) {
-            repository.deleteFile(id)
-            success = true
+    fun deleteFile(request: HttpServletRequest, @RequestParam id: String): Boolean {
+        if (isKeyValid(request)) {
+            val key = getKeyFromRequest(request)
+            var success = false
+            if (repository.findFileByKey(key).any { it.id.equals(id) }) {
+                repository.deleteFile(id)
+                success = true
+            }
+            return success
+        } else {
+            throw InvalidKeyException()
         }
-        return success
+
     }
 
     @CrossOrigin
@@ -104,7 +104,6 @@ class APIController {
 
         return File("meta.json").readText()
     }
-
 
 
     @CrossOrigin
@@ -141,7 +140,7 @@ class APIController {
 
     private fun ExecuteManifest(mf: Manifest): MutableList<Any> {
         var pipe = PipelineBuilder(mf)
-        pipe.buildPipeline(PipelineBuilder.AssetMechanism.All,repository)
+        pipe.buildPipeline(PipelineBuilder.AssetMechanism.All, repository)
         var outputs = pipe.executePipes()
         return outputs
     }
@@ -157,6 +156,7 @@ class APIController {
             return "DEVKEY"
         }
     }
+
     private fun isAuthorisedToSaveService(request: HttpServletRequest, space: String): Boolean {
         if (environment.getActiveProfiles().contains("prod")) {
             val AuthURI = Config.get("AuthURI")
@@ -178,6 +178,29 @@ class APIController {
             return authorisationRequest.text == "true"
         }
         return true
+    }
+
+    private fun isKeyValid(request: HttpServletRequest): Boolean {
+       // if (environment.getActiveProfiles().contains("prod")) {
+            val AuthURI = Config.get("AuthURI")
+
+            // http://www.baeldung.com/get-user-in-spring-security
+            val raw = request.getHeader("authorization")
+            if (raw == null) return false;
+            val apikey = String(Base64.getDecoder().decode(raw.removePrefix("Basic ")))
+
+            val user = apikey.split(":")[0]
+            val pass = apikey.split(":")[1]
+
+
+            val authorisationRequest = get(AuthURI + "api/checkKey",
+                    auth = BasicAuthorization(user, pass)
+            )
+
+            if (authorisationRequest.statusCode != 200) return false
+            return authorisationRequest.text == "true"
+        //}
+        //return true
     }
 
     data class Event(var key: String = "", var action: String = "", var type: String = "", var name: String = "", var reason: String = "", var content: String = "")
